+++ src/controllers/queries/index.ts
@@ -8,7 +8,8 @@ import { PartialQueryPayload } from '../../models/arrays.js'
 import Database from '../../models/db/index.js'
 import { ConnectionRow, QueryRow, Where } from '../../models/db/types.js'
 import { type UUID } from '../../models/strings.js'
-import VeritableCloudagent, { DrpcResponse } from '../../models/veritableCloudagent.js'
+import VeritableCloudagent from '../../models/veritableCloudagent/index.js'
+import { DrpcResponse } from '../../models/veritableCloudagentInt.js'
 import QueriesTemplates from '../../views/queries/queries.js'
 import QueryListTemplates from '../../views/queries/queriesList.js'
 import Scope3CarbonConsumptionTemplates from '../../views/queries/requestCo2scope3.js'
