+++ src/controllers/reset/index.ts
@@ -7,7 +7,8 @@ import { Env } from '../../env/index.js'
 import { BadRequestError, InternalError } from '../../errors.js'
 import Database from '../../models/db/index.js'
 import type { TABLE } from '../../models/db/types.js'
-import VeritableCloudagent, { Connection, Credential } from '../../models/veritableCloudagent.js'
+import VeritableCloudagent from '../../models/veritableCloudagent/index.js'
+import { Connection, Credential } from '../../models/veritableCloudagentInt.js'
 
 @injectable()
 @Security('oauth2')
