+++ src/init.ts
@@ -6,7 +6,7 @@ import { container } from 'tsyringe'
 import { InitEnv } from './env/index.js'
 import { type ILogger } from './logger.js'
 import { CredentialSchema } from './models/credentialSchema.js'
-import VeritableCloudagent from './models/veritableCloudagent.js'
+import VeritableCloudagent from './models/veritableCloudagent/index.js'
 
 const env = container.resolve(InitEnv)
