+++ src/models/veritableCloudagentInt.ts
@@ -1,9 +1,8 @@
-import { inject, injectable, singleton } from 'tsyringe'
 import { z } from 'zod'
 
-import { Env, type PartialEnv } from '../env/index.js'
+import { type PartialEnv } from '../env/index.js'
 import { InternalError } from '../errors.js'
-import { Logger, type ILogger } from '../logger.js'
+import { type ILogger } from '../logger.js'
 
 const oobParser = z.object({
   invitationUrl: z.string(),
@@ -159,12 +158,13 @@ export type CredentialProposalAcceptInput = {
 
 type parserFn<O> = (res: Response) => O | Promise<O>
 
-@singleton()
-@injectable()
-export default class VeritableCloudagent {
+/*
+  This is in internal class used for e2e tests. Rest of the repository is using VeritableCloudagent which extends this class.
+*/
+export class VeritableCloudagentInt {
   constructor(
-    @inject(Env) private env: PartialEnv<'CLOUDAGENT_ADMIN_ORIGIN'>,
-    @inject(Logger) protected logger: ILogger
+    private env: PartialEnv<'CLOUDAGENT_ADMIN_ORIGIN'>,
+    protected logger: ILogger
   ) {}
 
   public async createOutOfBandInvite(params: { companyName: string }): Promise<OutOfBandInvite> {
