+++ src/services/credentialEvents/companyDetailsV1.ts
@@ -5,11 +5,8 @@ import { Env } from '../../env/index.js'
 import { Logger, type ILogger } from '../../logger.js'
 import { CredentialSchema } from '../../models/credentialSchema.js'
 import Database from '../../models/db/index.js'
-import {
-  default as VeritableCloudagent,
-  type Credential,
-  type CredentialFormatData,
-} from '../../models/veritableCloudagent.js'
+import { default as VeritableCloudagent } from '../../models/veritableCloudagent/index.js'
+import { type Credential, type CredentialFormatData } from '../../models/veritableCloudagentInt.js'
 import { CredentialEventHandler } from './types.js'
 
 @singleton()
