+++ src/services/veritableCloudagentEvents.ts
@@ -4,7 +4,9 @@ import { z } from 'zod'
 
 import { Env } from '../env/index.js'
 import { Logger, type ILogger } from '../logger.js'
-import VeritableCloudagent, { connectionParser, credentialParser } from '../models/veritableCloudagent.js'
+import VeritableCloudagent from '../models/veritableCloudagent/index.js'
+import { connectionParser, credentialParser } from '../models/veritableCloudagentInt.js'
+
 import IndexedAsyncEventEmitter from '../utils/indexedAsyncEventEmitter.js'
 import { MapDiscriminatedUnion } from '../utils/types.js'
