+++ test/helpers/cloudagent.ts
@@ -1,7 +1,7 @@
 import { container } from 'tsyringe'
 
 import { Env } from '../../src/env/index.js'
-import VeritableCloudagent from '../../src/models/veritableCloudagent.js'
+import VeritableCloudagent from '../../src/models/veritableCloudagent/index.js'
 import { validCompanyName, validCompanyNumber } from './fixtures.js'
 import { mockLogger } from './logger.js'
