+++ test/integration/partialQueryAggregation.test.ts
@@ -3,7 +3,7 @@ import type express from 'express'
 import { afterEach, beforeEach, describe } from 'mocha'
 
 import Database from '../../src/models/db/index.js'
-import VeritableCloudagent from '../../src/models/veritableCloudagent.js'
+import VeritableCloudagent from '../../src/models/veritableCloudagent/index.js'
 import { cleanupCloudagent } from '../helpers/cloudagent.js'
 import { withCompanyHouseMock } from '../helpers/companyHouse.js'
 import { cleanup } from '../helpers/db.js'
