+++ test/integration/resetController.test.ts
@@ -8,7 +8,9 @@ import { cleanup } from '../helpers/db.js'
 
 import Database from '../../src/models/db/index.js'
 import { ConnectionRow } from '../../src/models/db/types.js'
-import VeritableCloudagent, { Connection } from '../../src/models/veritableCloudagent.js'
+import EmailService from '../../src/models/emailService/index.js'
+import VeritableCloudagent from '../../src/models/veritableCloudagent/index.js'
+import { Connection } from '../../src/models/veritableCloudagentInt.js'
 import createHttpServer from '../../src/server.js'
 import VeritableCloudagentEvents from '../../src/services/veritableCloudagentEvents.js'
 import { withCompanyHouseMock } from '../helpers/companyHouse.js'
@@ -18,6 +20,7 @@ import { del } from '../helpers/routeHelper.js'
 describe('integration test for /reset endpoint', function () {
   const db = container.resolve(Database)
   const cloudagent = container.resolve(VeritableCloudagent)
+  const email = container.resolve(EmailService)
 
   afterEach(async () => {
     await cleanup()
@@ -51,7 +54,7 @@ describe('integration test for /reset endpoint', function () {
       context.cloudagentEvents.stop()
     })
 
-    withVerifiedConnection(context)
+    withVerifiedConnection(context, email)
 
     it('removes all connections and credentials and returns 200', async function () {
       const localConnections: ConnectionRow[] = await db.get('connection', {})
